rm deprecated code
