+++ tests/integration/util_query.go
@@ -75,26 +75,6 @@ func (s *MiniClusterSuite) waitForLoadInternal(ctx context.Context, dbName, coll
 	}
 }
 
-func waitingForLoad(ctx context.Context, cluster *MiniCluster, collection string) {
-	getLoadingProgress := func() *milvuspb.GetLoadingProgressResponse {
-		loadProgress, err := cluster.Proxy.GetLoadingProgress(ctx, &milvuspb.GetLoadingProgressRequest{
-			CollectionName: collection,
-		})
-		if err != nil {
-			panic("GetLoadingProgress fail")
-		}
-		return loadProgress
-	}
-	for getLoadingProgress().GetProgress() != 100 {
-		select {
-		case <-ctx.Done():
-			panic("load timeout")
-		default:
-			time.Sleep(500 * time.Millisecond)
-		}
-	}
-}
-
 func ConstructSearchRequest(
 	dbName, collectionName string,
 	expr string,
